refactor: move cli code to the separate function
